etl_flow/pyproject.toml
[tool.poetry]
name = "etl-flow"
version = "0.1.0"
description = "ETL Flow"
authors = ["FernandoCelmer <email@fernandocelmer.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.3"
pydantic = "^2.10.6"
dotflow = "0.13.0.dev1"
pytest = "^8.3.5"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
